# frontend/src/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hotdeal List API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuthentication": []
    }
  ],
  "tags": [
    {
      "name": "Category",
      "description": "카테고리 API"
    },
    {
      "name": "Platform",
      "description": "플랫폼 API"
    },
    {
      "name": "Hotdeal",
      "description": "핫딜 조회 API"
    }
  ],
  "paths": {
    "/v1/platforms": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "플랫폼 목록 조회",
        "description": "지원하는 플랫폼(커뮤니티) 목록을 반환",
        "operationId": "listPlatforms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/hotdeals": {
      "get": {
        "tags": [
          "Hotdeal"
        ],
        "summary": "핫딜 목록 조회",
        "description": "필터, 검색, 정렬을 지원하는 핫딜 목록 API",
        "operationId": "listHotdeals",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "페이지네이션 커서",
            "required": false,
            "schema": {
              "type": "string",
              "description": "페이지네이션 커서"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "페이지 크기 (최대 100)",
            "required": false,
            "schema": {
              "maximum": 100,
              "type": "integer",
              "description": "페이지 크기 (최대 100)",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "정렬 기준",
            "required": false,
            "schema": {
              "type": "string",
              "description": "정렬 기준",
              "default": "LATEST",
              "enum": [
                "LATEST",
                "POPULAR",
                "COMMENTS",
                "VIEWS"
              ]
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "검색 키워드",
            "required": false,
            "schema": {
              "type": "string",
              "description": "검색 키워드"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "description": "카테고리 코드 목록",
            "required": false,
            "schema": {
              "type": "array",
              "description": "카테고리 코드 목록",
              "items": {
                "type": "string",
                "description": "카테고리 코드 목록"
              }
            }
          },
          {
            "name": "platforms",
            "in": "query",
            "description": "플랫폼 목록",
            "required": false,
            "schema": {
              "type": "array",
              "description": "플랫폼 목록",
              "items": {
                "type": "string",
                "description": "플랫폼 목록",
                "enum": [
                  "COOLENJOY_JIRUM",
                  "QUASARZONE_JIRUM",
                  "QUASARZONE_TASEYO",
                  "RULIWEB_HOTDEAL",
                  "PPOMPPU_PPOMPPU",
                  "PPOMPPU_HOTDEAL",
                  "PPOMPPU_OVERSEAS_HOTDEAL",
                  "CLIEN_ALTTEUL"
                ]
              }
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "최소 가격",
            "required": false,
            "schema": {
              "type": "number",
              "description": "최소 가격"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "최대 가격",
            "required": false,
            "schema": {
              "type": "number",
              "description": "최대 가격"
            }
          },
          {
            "name": "includeEnded",
            "in": "query",
            "description": "종료된 핫딜 포함 여부",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "종료된 핫딜 포함 여부",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HotdealListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hotdeals/{id}": {
      "get": {
        "tags": [
          "Hotdeal"
        ],
        "summary": "핫딜 상세 조회",
        "operationId": "getHotdeal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "핫딜 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HotdealResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hotdeals/{id}/price-history": {
      "get": {
        "tags": [
          "Hotdeal"
        ],
        "summary": "가격 히스토리 조회",
        "description": "유사 상품명 기반 날짜별 가격 추이 조회",
        "operationId": "getPriceHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "핫딜 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PriceHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hotdeals/suggest": {
      "get": {
        "tags": [
          "Hotdeal"
        ],
        "summary": "검색어 자동완성",
        "description": "입력된 키워드로 상품명 자동완성 추천",
        "operationId": "suggest",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "검색 키워드",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "추천 개수 (기본 7)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "카테고리 목록 조회",
        "description": "트리 구조의 카테고리 목록을 반환",
        "operationId": "listCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PlatformResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "플랫폼 코드",
            "example": "COOLENJOY_JIRUM"
          },
          "communityName": {
            "type": "string",
            "description": "커뮤니티(사이트)명",
            "example": "쿨엔조이"
          },
          "displayName": {
            "type": "string",
            "description": "플랫폼 표시명",
            "example": "쿨엔조이 지름"
          }
        },
        "description": "플랫폼 정보"
      },
      "HotdealListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotdealResponse"
            }
          },
          "nextCursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "HotdealResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "platformType": {
            "type": "string",
            "enum": [
              "COOLENJOY_JIRUM",
              "QUASARZONE_JIRUM",
              "QUASARZONE_TASEYO",
              "RULIWEB_HOTDEAL",
              "PPOMPPU_PPOMPPU",
              "PPOMPPU_HOTDEAL",
              "PPOMPPU_OVERSEAS_HOTDEAL",
              "CLIEN_ALTTEUL"
            ]
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currencyUnit": {
            "type": "string"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "commentCount": {
            "type": "integer",
            "format": "int32"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "shoppingPlatform": {
            "type": "string"
          },
          "categoryCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wroteAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "highlightedTitle": {
            "type": "string"
          },
          "highlightedProductName": {
            "type": "string"
          },
          "ended": {
            "type": "boolean"
          }
        }
      },
      "DailyPriceStats": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "날짜 (yyyy-MM-dd)",
            "format": "date"
          },
          "count": {
            "type": "integer",
            "description": "해당 날짜의 핫딜 수",
            "format": "int32"
          },
          "minPrice": {
            "type": "number",
            "description": "최저 가격"
          },
          "maxPrice": {
            "type": "number",
            "description": "최고 가격"
          },
          "avgPrice": {
            "type": "number",
            "description": "평균 가격"
          },
          "hotdeals": {
            "type": "array",
            "description": "해당 날짜의 핫딜 목록",
            "items": {
              "$ref": "#/components/schemas/HotdealSummary"
            }
          }
        },
        "description": "일별 가격 통계"
      },
      "HotdealSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "핫딜 ID",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "게시글 제목"
          },
          "productName": {
            "type": "string",
            "description": "상품명"
          },
          "price": {
            "type": "number",
            "description": "가격"
          },
          "url": {
            "type": "string",
            "description": "원본 게시글 URL"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "썸네일 URL"
          }
        },
        "description": "핫딜 요약 정보"
      },
      "PriceHistoryResponse": {
        "type": "object",
        "properties": {
          "hotdealId": {
            "type": "integer",
            "description": "기준 핫딜 ID",
            "format": "int64"
          },
          "productName": {
            "type": "string",
            "description": "기준 핫딜 상품명"
          },
          "totalSimilarCount": {
            "type": "integer",
            "description": "유사 상품 총 개수",
            "format": "int64"
          },
          "priceHistory": {
            "type": "array",
            "description": "날짜별 가격 통계 목록",
            "items": {
              "$ref": "#/components/schemas/DailyPriceStats"
            }
          }
        },
        "description": "가격 히스토리 응답"
      },
      "SuggestResponse": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuthentication": {
        "type": "http",
        "description": "JWT 액세스 토큰을 입력하세요",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}